Sprint1/formu nouveau titre

+++ src/main/java/com/mada/visa_mg/repository/DossierRepository.java
@@ -1,8 +1,7 @@
 package com.mada.visa_mg.repository;
 
+import com.mada.visa_mg.entity.Dossier;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.mada.visa_mg.domain.entity.Dossier;
-
-public interface DossierRepository extends JpaRepository<Dossier, Long> {
+public interface DossierRepository extends JpaRepository<Dossier, Integer> {
 }
